+++ RELEASENOTES.md
@@ -54,6 +54,7 @@ The following binary downloads have been compiled with ESP8266/Arduino library c
 
 ### Version 8.3.1.7
 
+- Remove Arduino ESP8266 Core support for versions before 2.7.1
 - Change to limited support of Arduino IDE as an increasing amount of features cannot be compiled with Arduino IDE
 - Change IRremoteESP8266 library from v2.7.6 to v2.7.8
 - Change Adafruit_SGP30 library from v1.0.3 to v1.2.0 (#8519)